+++ src/lib.rs
@@ -1431,6 +1431,7 @@ impl WebView {
         clipboard = false,
         javascript_enabled = true,
         autoplay = true,
+        hotkeys_zoom = false,
         focused = true,
         background_color = None,
         user_agent = None,
@@ -1466,6 +1467,7 @@ impl WebView {
         clipboard: bool,
         javascript_enabled: bool,
         autoplay: bool,
+        hotkeys_zoom: bool,
         focused: bool,
         background_color: Option<(u8, u8, u8, u8)>,
         user_agent: Option<String>,
@@ -1821,6 +1823,7 @@ WebViews that share a session must use the same app= root \
             builder = builder.with_javascript_disabled();
         }
         builder = builder.with_autoplay(autoplay);
+        builder = builder.with_hotkeys_zoom(hotkeys_zoom);
         if has_permission_handler {
             let permission_cb_clone = permission_cb.clone();
             let permission_sync_pending_clone = permission_sync_pending.clone();